SAC: rename train_SAC.py → sac.py, extract feature builders
* git-mv ``src/ajax/agents/SAC/train_SAC.py`` → ``sac.py`` (same content
  initially — the rename is the structural change; behaviour identical).
* Extract the 14-feature hook-builder block (``make_action_pipeline``,
  ``make_target_modifier``, ``make_policy_obs_preprocessor``,
  ``make_policy_action_transform``, ``make_bc_loss_fn``,
  ``make_runtime_maintenance``, ``make_eval_action_transform``,
  ``make_next_expert_fn``, ``ActionPipelineResult``, ``_apply_lcb_gate``)
  out of ``sac.py`` into ``src/ajax/extensions/_sac_hooks.py`` — moved
  verbatim, so every feature's code path is byte-identical to before.
* Update REDQ's lineage import (4 symbols) + the two test importers +
  ``SAC.SAC``'s ``make_train`` import to the new module path.
* Add thin :class:`Extension` config classes for the 14 features in
  ``src/ajax/extensions/{expert,target_mods,exploration,pretrain}.py``
  (no behaviour yet — they are wired by the next commit).

Plain-SAC behaviour-parity gate passes; REDQ smoke tests pass.
